Set grace termination period of testrunners pod to 0
When running multiple tests, a race condition might happen between
them: When a test finishes, it deletes the testrunner pod. When a
test starts, it creates a testrunner pod. Sometimes this creation
fails because the testrunner pod from previous test was not entirely
removed. It is in process of deletion.

By setting the graceful shutdown period to 0 we make sure the testrunner
pod will be removed immediately, thus elimiting this race condition.