Drop the usage of "router.hostIP" system property
We already have the property "openshift.router.host" with the
same purpose. Better yet, it can be replaced by an env variable
"OPENSHIFT_ROUTER_HOST". It's up to the user to choose between
them.

Having another system property might lead to some confusion.

We should also update our doc (currently README's) to reflect
that. As they are outdated, it makes sense to fix them along
with the other necessary changes. An issue is open for that
(https://github.com/jboss-openshift/ce-testsuite/issues/138)